pyproject.toml
[project]
name = "notion-cli"
version = "0.1.0"
description = "CLI tool for Notion workspace management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "notion-client>=2.7.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
    "typer>=0.21.1",
]

[project.scripts]
notion = "notion_cli.main:app"
